# desktop_app/openapi/archestra/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Archestra API",
    "version": "0.0.1"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/chat": {
      "get": {
        "operationId": "getChats",
        "tags": ["Chat"],
        "description": "Get all chats",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createChat",
        "tags": ["Chat"],
        "description": "Create new chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/chat/{id}": {
      "get": {
        "operationId": "getChatById",
        "tags": ["Chat"],
        "description": "Get single chat with messages",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateChat",
        "tags": ["Chat"],
        "description": "Update chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteChat",
        "tags": ["Chat"],
        "description": "Delete chat",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/external_mcp_client": {
      "get": {
        "operationId": "getConnectedExternalMcpClients",
        "tags": ["External MCP Client"],
        "description": "Get all connected external MCP clients",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/external_mcp_client/supported": {
      "get": {
        "operationId": "getSupportedExternalMcpClients",
        "tags": ["External MCP Client"],
        "description": "Get supported external MCP client names",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/external_mcp_client/connect": {
      "post": {
        "operationId": "connectExternalMcpClient",
        "tags": ["External MCP Client"],
        "description": "Connect an external MCP client",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/external_mcp_client/{client_name}/disconnect": {
      "delete": {
        "operationId": "disconnectExternalMcpClient",
        "tags": ["External MCP Client"],
        "description": "Disconnect an external MCP client",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "client_name",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/llm/stream": {
      "post": {
        "operationId": "streamLlmResponse",
        "tags": ["LLM"],
        "description": "Stream LLM response",
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/mcp_request_log": {
      "get": {
        "operationId": "getMcpRequestLogs",
        "tags": ["MCP Request Log"],
        "description": "Get MCP request logs with filtering and pagination",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "clearMcpRequestLogs",
        "tags": ["MCP Request Log"],
        "description": "Clear MCP request logs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_request_log/{id}": {
      "get": {
        "operationId": "getMcpRequestLogById",
        "tags": ["MCP Request Log"],
        "description": "Get a single MCP request log by ID",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_request_log/stats": {
      "get": {
        "operationId": "getMcpRequestLogStats",
        "tags": ["MCP Request Log"],
        "description": "Get MCP request log statistics",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_server": {
      "get": {
        "operationId": "getMcpServers",
        "tags": ["MCP Server"],
        "description": "Get all installed MCP servers",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_server/install": {
      "post": {
        "operationId": "installMcpServer",
        "tags": ["MCP Server"],
        "description": "Install MCP server from catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_server/install_custom": {
      "post": {
        "operationId": "installCustomMcpServer",
        "tags": ["MCP Server"],
        "description": "Install custom MCP server",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_server/{slug}": {
      "delete": {
        "operationId": "uninstallMcpServer",
        "tags": ["MCP Server"],
        "description": "Uninstall MCP server",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/mcp_server/start_oauth": {
      "post": {
        "operationId": "startMcpServerOauth",
        "tags": ["MCP Server"],
        "description": "Start MCP server OAuth flow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/sandbox/status": {
      "get": {
        "operationId": "getSandboxStatus",
        "tags": ["Sandbox"],
        "description": "Get the current status of the sandbox environment",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  }
}